+++ netflow-ksp/src/main/kotlin/com/kmpbits/netflow_ksp/Fqns.kt
@@ -1,0 +1,27 @@
+package com.kmpbits.netflow_ksp
+
+internal object Fqns {
+    const val NET_FLOW_API = "com.kmpbits.netflow_annotations.NetFlowApi"
+    const val GET = "com.kmpbits.netflow_annotations.GET"
+    const val POST = "com.kmpbits.netflow_annotations.POST"
+    const val PUT = "com.kmpbits.netflow_annotations.PUT"
+    const val DELETE = "com.kmpbits.netflow_annotations.DELETE"
+    const val PATCH = "com.kmpbits.netflow_annotations.PATCH"
+    const val PATH = "com.kmpbits.netflow_annotations.Path"
+    const val QUERY = "com.kmpbits.netflow_annotations.Query"
+    const val HEADER = "com.kmpbits.netflow_annotations.Header"
+    const val BODY = "com.kmpbits.netflow_annotations.Body"
+
+    val HTTP_METHOD_ANNOTATIONS = setOf(GET, POST, PUT, DELETE, PATCH)
+
+    const val FLOW = "kotlinx.coroutines.flow.Flow"
+    const val RESULT_STATE = "com.kmpbits.netflow_core.states.ResultState"
+    const val ASYNC_STATE = "com.kmpbits.netflow_core.states.AsyncState"
+    const val LIST = "kotlin.collections.List"
+    const val MAP = "kotlin.collections.Map"
+    const val STRING = "kotlin.String"
+
+    const val CLIENT_PKG = "com.kmpbits.netflow_core.client"
+    const val ENUMS_PKG = "com.kmpbits.netflow_core.enums"
+    const val DESERIALIZABLES_PKG = "com.kmpbits.netflow_core.deserializables"
+}